+++ tdrs-backend/tdpservice/users/change_request_signals.py
@@ -2,9 +2,6 @@
 
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from django.core.mail import send_mail
-from django.conf import settings
-from django.utils.translation import gettext_lazy as _
 
 from .models import UserChangeRequest, UserChangeRequestStatus
